Moved all the structures to structs folder.
